+++ epyk/core/html/HtmlOthers.py
@@ -627,7 +627,8 @@ class HtmlQRCode(Html.Html):
     return "%s_obj" % self.htmlCode
 
   _js__builder__ = ''' htmlObj.innerHTML = ""; 
-       window[ htmlObj.id + '_obj'] = new QRCode(htmlObj, Object.assign({'text': data}, options) )'''
+       window[ htmlObj.id + '_obj'] = new QRCode(htmlObj, Object.assign({'text': data}, options) );
+       window[ htmlObj.id + '_obj']._el.querySelector("img").style["margin"] = 0'''
 
   @property
   def js(self):